// src/app/BusinessModel/Basketball/basketball-enum.ts
export enum BasketballEnum {
    Date = 'Date',
    ContestGroupId = 'ContestGroupId',
    ContestGroupText = 'ContestGroupText',
    RoundName = 'RoundName',
    RoundId = 'RoundId',
    StatsTypeId = 'StatsTypeId',
    StatsTypeValue = 'StatsTypeValue',
    CountryId = 'CountryId',
    CountryName = 'CountryName',
    SelectedHomeTeamId = 'SelectedHomeTeamId',
    SelectedAwayTeamId = 'selectedAwayTeamId',
    SelectedHomeTeamName = 'SelectedHomeTeamName',
    SelectedAwayTeamName = 'selectedAwayTeamName',
}